added/requirements.txt
# Core Framework
eel>=0.14.0
python-dotenv>=1.0.0

# AI and ML
openai-whisper
pyttsx3>=2.90
cohere>=4.0.0
groq>=0.4.0
requests>=2.25.0

# Database
pymongo>=4.0.0

# Computer Vision and Authentication
opencv-python>=4.5.0
face-recognition>=1.3.0
Pillow>=8.0.0

# Audio Processing
pyaudio>=0.2.11
sounddevice>=0.4.0
soundfile>=0.12.0
playsound>=1.2.0
SpeechRecognition>=3.8.0

# System Integration
pyautogui>=0.9.50
pywhatkit>=5.0
keyboard>=0.13.0
psutil>=5.8.0

# Web and HTTP
fastapi>=0.68.0
uvicorn>=0.15.0

# Google API
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
certifi>=2021.0.0

# Utilities
rich>=10.0.0
beautifulsoup4>=4.9.0
numpy>=1.21.0
torch>=1.9.0
torchaudio>=0.9.0